Annotate Jccs ending DSB blocks
In my testing, this has now several times been the cause
for otherwise hard-to-understand DSB->MITE switches on
SKL derivatives, especially with the "both halves of the
64B cache line need DSB entries" rule, so it seems worth
pointing out the relevant branches in the output.

This involves keeping track of addresses when setting up the
TableLineData and checking whether branches or fused pairs
fulfill the conditions for the Jcc erratum.