Enable disassembling hex strings
This patch enables disassembling hex strings by changing the input
options to the uiCA.py script. Instead of accepting a file as a
positional argument, a file is now accepted as a flag. Input can also be
accepted with the hex flag, assuming a hex string is passed in. This
allows for more easily working with the BHive dataset using the main
uiCA.py script in addition to other tools that pass around various
representations of machine code without having to write anything to a
file.